% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_function_arguments.R
\name{check_colorpalette}
\alias{check_colorpalette}
\title{Establish the default behavior for colorpalette}
\usage{
check_colorpalette(shareddata, colorvar = NULL, colorpalette = NULL)
}
\arguments{
\item{shareddata}{The SharedData object provided to the calling function.}

\item{colorvar}{The \code{colorvar} argument provided to the calling function.}

\item{colorpalette}{The \code{colorpalette} argument provided to the calling
function.}
}
\value{
The color palette to be used for the visualization.
}
\description{
Helper function for internal use only.
}
\keyword{internal}
