% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_function_arguments.R
\name{check_colorvar}
\alias{check_colorvar}
\title{Check the colorvar argument for \code{driveplot_companion()}
and \code{driveplot_map()}}
\usage{
check_colorvar(shareddata, colorvar)
}
\arguments{
\item{shareddata}{The SharedData object provided to the calling function.}

\item{colorvar}{The \code{colorvar} argument provided to the calling function.}
}
\value{
A value indicating whether colorvar is not a column in the data
(NULL), colorvar is a numeric column in the data (TRUE), or colorvar is
not a numeric column in the data (FALSE).
}
\description{
Helper function for internal use only.
}
\keyword{internal}
