% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_function_arguments.R
\name{check_function_arguments}
\alias{check_function_arguments}
\title{Check the arguments provided to a function and throw an error if there is
a problem.}
\usage{
check_function_arguments(
  shareddata,
  checks = c("x", "y", "colorvar", "lat", "lng"),
  x = NULL,
  y = NULL,
  colorvar = NULL,
  lng = NULL,
  lat = NULL
)
}
\arguments{
\item{shareddata}{The SharedData object provided to the calling function.}

\item{checks}{A vector of the checks to be performed; defaults to
\code{c("x", "y", "colorvar", "lat", "lng")}}

\item{x}{The \code{x} argument provided to the calling function.}

\item{y}{The \code{y} argument provided to the calling function.}

\item{colorvar}{The \code{colorvar} argument provided to the calling function.}

\item{lng}{The \code{lng} argument provided to the calling function.}

\item{lat}{The \code{lat} argument provided to the calling function.}
}
\value{
TRUE if all function arguments pass the checks.
}
\description{
Helper function for internal use only.
}
