% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_function_arguments.R
\name{check_lnglat}
\alias{check_lnglat}
\title{Check the latitude and longitude for \code{driveplot_map()}
Helper function for internal use only.}
\usage{
check_lnglat(shareddata, lng, lat)
}
\arguments{
\item{shareddata}{The SharedData object provided to the calling function.}

\item{lng}{The \code{lng} argument provided to the calling function.}

\item{lat}{The \code{lat} argument provided to the calling function.}
}
\value{
If \code{lng} and \code{lat} are correctly provided, return NULL. If an sf
geometry column is provided, return a dataframe of the points derived from
the geometry column.
}
\description{
Check the latitude and longitude for \code{driveplot_map()}
Helper function for internal use only.
}
\keyword{internal}
