% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_function_arguments.R
\name{check_xy}
\alias{check_xy}
\title{Check the arguments provided to \code{driveplot_companion()}, \code{driveplot_map()},
and \code{driveplot_map()}
Function to check the x and y arguments for \code{driveplot_companion()}.}
\usage{
check_xy(shareddata, arg, argname)
}
\arguments{
\item{shareddata}{The SharedData object provided to the calling function.}

\item{arg}{The \code{x} or \code{y} argument provided to the calling function.}

\item{argname}{The name (x or y) of the argument to be checked.}
}
\value{
The result of the argument tidily evaluated in the environment
of the original data from \code{shareddata}.
}
\description{
Helper function for internal use only.
}
\keyword{internal}
