% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_function_arguments.R
\name{check_ylabels}
\alias{check_ylabels}
\title{Check the \code{ylabels} argument for \code{driveplot_companions()}
Helper function for internal use only.}
\usage{
check_ylabels(ylabels, ylength)
}
\arguments{
\item{ylabels}{A vector or list of labels provided to the calling function.}

\item{ylength}{The length of the ys argument from the calling function.}
}
\value{
Either \code{ylabels} or a list of NULL of length \code{ylength}.
}
\description{
Check the \code{ylabels} argument for \code{driveplot_companions()}
Helper function for internal use only.
}
\keyword{internal}
