% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_SharedData.R
\name{convert_to_SharedData}
\alias{convert_to_SharedData}
\title{Convert a data frame to a SharedData object and warn the user
this was done}
\usage{
convert_to_SharedData(df)
}
\arguments{
\item{df}{The provided data object to be converted to a SharedData object.}
}
\value{
A SharedData object.
}
\description{
Helper function for internal use only.
}
\keyword{internal}
