% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driveplot_companion.R
\name{driveplot_companion}
\alias{driveplot_companion}
\title{Make a single plotly scatter plot from shared drive data}
\usage{
driveplot_companion(
  shareddata,
  x,
  y,
  colorvar = NULL,
  colorpalette = "#03F",
  xlab = NULL,
  ylab = NULL,
  showlegend = TRUE,
  legendtitle = NULL
)
}
\arguments{
\item{shareddata}{A SharedData object containing observations to be plotted.}

\item{x}{The variable from shareddata to be plotted on the horizontal axis.}

\item{y}{The variable from shareddata to be plotted on the vertical axis.}

\item{colorvar}{The variable in shareddata to which color should be mapped.}

\item{colorpalette}{The color palette for the plot; either a single
color (e.g., "red") or one of "viridis", "inferno", "magma", or "plasma".}

\item{xlab}{The label for the variable on the horizontal axis.}

\item{ylab}{The label for the variable on the vertical axis.}

\item{showlegend}{Show the plot legend (TRUE) or not (FALSE).}

\item{legendtitle}{The title for the plot legend.}
}
\value{
A plotly scatterplot.
}
\description{
Make a single plotly scatter plot from shared drive data
}
\examples{
library(crosstalk)
library(dplyr)
data(nds_data)
nds_sf7 <- nds_data |>
  filter(drive==7) |>
  sf::st_as_sf(coords = c("gps_long", "gps_lat"), crs = "WGS84")
nds_sf7_sd <- SharedData$new(nds_sf7)

# Time series of speed
driveplot_companion(nds_sf7_sd, time_cst, speed_mph)

# Color points by direction of car
driveplot_companion(shareddata = nds_sf7_sd,
                    x = time_cst,
                    y = speed_mph,
                    colorvar = gyro_heading,
                    colorpalette = "viridis")
}
