% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driveplot_map.R
\name{driveplot_map}
\alias{driveplot_map}
\title{Create a standalone map}
\usage{
driveplot_map(
  shareddata,
  lng = NULL,
  lat = NULL,
  colorvar = NULL,
  label = NA,
  colorpalette = NULL,
  fillopacity = 1,
  mapheight = "100vh",
  tileprovider = "OpenStreetMap"
)
}
\arguments{
\item{shareddata}{A SharedData object containing observations to be plotted.}

\item{lng}{The bare (unquoted) column of shareddata containing longitude
(only required if shareddata does not have a geometry column).}

\item{lat}{The bare (unquoted) column of shareddata containing latitude
(only required if shareddata does not have a geometry column).}

\item{colorvar}{The bare (unquoted) column in shareddata to which color
should be mapped.}

\item{label}{An optional label for the map points.}

\item{colorpalette}{The color palette for the plot; either a single
color (e.g., "red") or one of the viridis color palettes compatible
with leaflet. Run \code{leaflet_color_palettes()} to see the available options
for viridis color palettes. If \code{colorvar} and \code{colorpalette} are both not
provided, \code{colorpalette} defaults to blue. If \code{colorvar} is provided but
\code{colorpalette} is not, \code{colorpalette} defaults to the color palette
"viridis".}

\item{fillopacity}{The opacity of the fill of the map points (0 to 1).}

\item{mapheight}{The height of the map, e.g, "100vh" or 400px".}

\item{tileprovider}{The name of the provider for the basemap tiles. Options
can be found by running \code{leaflet::providers}. Defaults to "OpenStreetMap".}
}
\value{
A leaflet map.
}
\description{
Create a standalone map
}
\examples{
library(crosstalk)
data(drive7)
shared_drive <- SharedData$new(drive7)

# Basic map of one drive
driveplot_map(shareddata = shared_drive)

# Color drive points by direction of car
driveplot_map(
  shareddata = shared_drive,
  colorvar = gyro_heading,
  label = gyro_heading,
  colorpalette = "viridis"
)
}
