% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driveplot_map_color.R
\name{driveplot_map_color}
\alias{driveplot_map_color}
\title{Create the color palette to be used in \code{driveplot_map()}.}
\usage{
driveplot_map_color(colorvarnumeric, colorpalette, quocolor, ogdata)
}
\arguments{
\item{colorvarnumeric}{An indication of whether the color variable is
numeric (TRUE), not numeric (FALSE), or not provided (NULL)}

\item{colorpalette}{The color palette to use; either a single color
(e.g., "red") or one of "viridis", "inferno", "magma", or "plasma".}

\item{quocolor}{A quosure containing the color variable.}

\item{ogdata}{The data used to create the domain for the color palette.}
}
\value{
The leaflet color palette to be used in \code{driveplot_map()}.
}
\description{
Helper function for internal use only
}
