% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_color_palettes.R
\name{leaflet_color_palettes}
\alias{leaflet_color_palettes}
\title{Vector of viridis color palettes supported by Leaflet.}
\usage{
leaflet_color_palettes()
}
\value{
Vector of viridis color palettes supported by Leaflet.
}
\description{
Helper function
The map uses the same color palette as the companion graphs.
}
