% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driveplot.R
\name{driveplot}
\alias{driveplot}
\title{Create a vehicle trajectory plot map}
\usage{
driveplot(
  shareddata,
  lng = NULL,
  lat = NULL,
  x,
  ys,
  colorvar = NULL,
  maplabel = NA,
  colorpalette = NULL,
  fillopacity = 1,
  xlabel = NULL,
  ylabels = NULL,
  showlegend = TRUE,
  legendtitle = NULL,
  plottitle = NULL,
  spacing = 0.05,
  width = "100\%",
  height = "100vh"
)
}
\arguments{
\item{shareddata}{A SharedData object containing observations to be plotted.}

\item{lng}{The bare (unquoted) column of shareddata containing longitude
(only required if shareddata does not have a geometry column).}

\item{lat}{The bare (unquoted) column of shareddata containing latitude
(only required if shareddata does not have a geometry column)}

\item{x}{The bare (unquoted) column from shareddata to be plotted
on the horizontal axis.}

\item{ys}{A vector or list of bare (unquoted) columns from shareddata to be
plotted on the vertical axes of the companion graphs.}

\item{colorvar}{The bare (unquoted) column in shareddata to which color
should be mapped.}

\item{maplabel}{An optional label for the map points.}

\item{colorpalette}{The color palette for the plot; either a single
color (e.g., "red") or one of the viridis color palettes compatible
with leaflet. Run \code{leaflet_color_palettes()} to see the available options
for viridis color palettes. If \code{colorvar} and \code{colorpalette} are both not
provided, \code{colorpalette} defaults to blue. If \code{colorvar} is provided but
\code{colorpalette} is not, \code{colorpalette} defaults to the color palette
"viridis".}

\item{fillopacity}{The opacity of the fill of the map points (0 to 1).}

\item{xlabel}{The label for the variable on the horizontal axis.}

\item{ylabels}{A vector or list of labels for the vertical axes
of the companion graphs. If provided, it should be the same length as ys.}

\item{showlegend}{Show the plot legend (TRUE) or not (FALSE).}

\item{legendtitle}{The title for the plot legend.}

\item{plottitle}{The title for the plot map.}

\item{spacing}{A value between 0 and 1 for the space between the
companion graphs.}

\item{width}{The width of the plot map, e.g. "100\%" or "400px"}

\item{height}{The height of the plot map, e.g., "100vh" or 400px".}
}
\value{
A linked plot map.
}
\description{
Create a vehicle trajectory plot map
}
\examples{
library(crosstalk)
data(drive7)
shared_drive <- SharedData$new(drive7)

driveplot(
  shareddata = shared_drive,
  x = time_cst,
  ys = c(speed_mph, gyro_heading, gps_heading),
  colorvar = gyro_heading,
  maplabel = time_cst,
  colorpalette = "viridis",
  fillopacity = 1,
  xlabel = "Time",
  ylabels = c(
    "Speed (mph)", "Gyro Heading (degrees)",
    "GPS Heading (degrees)"
  ),
  showlegend = TRUE,
  legendtitle = "Gyro Heading",
  plottitle = "A drive with points colored by gyro heading",
  height = "500px"
)
}
