% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driveplot.R
\name{driveplot}
\alias{driveplot}
\title{Create a vehicle trajectory plot map}
\usage{
driveplot(
  shareddata,
  lng = NULL,
  lat = NULL,
  x,
  ys,
  colorvar = NULL,
  maplabel = NA,
  colorpalette = "#03F",
  fillOpacity = 1,
  xlabel = NULL,
  ylabels = NULL,
  showlegend = TRUE,
  legendtitle = NULL,
  plottitle = NULL,
  spacing = 0.05
)
}
\arguments{
\item{shareddata}{A SharedData object containing observations to be plotted.}

\item{lng}{The column of shareddata containing longitude
(only required if shareddata does not have a geometry column).}

\item{lat}{The column of shareddata containing latitude
(only required if shareddata does not have a geometry column)}

\item{x}{The variable from shareddata to be plotted on the horizontal axis.}

\item{ys}{A vector or list of 1 to 4 variables from shareddata to be
plotted on the vertical axes of the companion graphs.}

\item{colorvar}{The variable in shareddata to which color should be mapped.}

\item{maplabel}{An optional label for the map points.}

\item{colorpalette}{The color palette for the plot map; either a single
color (e.g., "red") or one of "viridis", "inferno", "magma", or "plasma".}

\item{fillOpacity}{The opacity of the fill of the map points (0 to 1).}

\item{xlabel}{The label for the variable on the horizontal axis.}

\item{ylabels}{A vector or list of 1 to 4 labels for the vertical axes
of the companion graphs.}

\item{showlegend}{Show the plot legend (TRUE) or not (FALSE).}

\item{legendtitle}{The title for the plot legend.}

\item{plottitle}{The title for the plot map.}

\item{spacing}{A value between 0 and 1 for the space between the
companion graphs.}
}
\value{
A linked plot map.
}
\description{
Create a vehicle trajectory plot map
}
\examples{
library(crosstalk)
data(drive7)
shared_drive <- SharedData$new(drive7)

driveplot(shareddata = shared_drive,
          x = time_cst,
          ys = c(speed_mph, gyro_heading, gps_heading),
          colorvar = gyro_heading,
          maplabel = time_cst,
          colorpalette = "viridis",
          fillOpacity = 1,
          xlabel = "Time",
          ylabels = c("Speed (mph)", "Gyro Heading (degrees)",
                      "GPS Heading (degrees)"))
}
