% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driveplot.R
\name{driveplot}
\alias{driveplot}
\title{Create a vehicle trajectory plot map}
\usage{
driveplot(
  shareddata,
  lng = NULL,
  lat = NULL,
  x,
  y1,
  y2 = NULL,
  y3 = NULL,
  y4 = NULL,
  colorvar = NULL,
  maplabel = NA,
  colorpalette = "#03F",
  fillOpacity = 1,
  xlabel = NULL,
  y1label = NULL,
  y2label = NULL,
  y3label = NULL,
  y4label = NULL,
  showlegend = TRUE,
  legendtitle = NULL,
  plottitle = NULL,
  spacing = 0.05
)
}
\arguments{
\item{shareddata}{A SharedData object containing observations to be plotted.}

\item{lng}{The column of shareddata containing longitude.
(only required if shareddata does not have a geometry column).}

\item{lat}{The column of shareddata containing latitude
(only required if shareddata does not have a geometry column)}

\item{x}{The variable from shareddata to be plotted on the horizontal axis.}

\item{y1}{The variable from shareddata to be plotted on the vertical axis
for the first graph.}

\item{y2}{The variable from shareddata to be plotted on the vertical axis
for the second graph.}

\item{y3}{The variable from shareddata to be plotted on the vertical axis
for the third graph.}

\item{y4}{The variable from shareddata to be plotted on the vertical axis
for the fourth graph.}

\item{colorvar}{The variable in shareddata to which color should be mapped.}

\item{maplabel}{An optional label for the map points.}

\item{colorpalette}{Color palette for the plot map; either a single color
(e.g., "red") or one of "viridis", "inferno", "magma", or "plasma".}

\item{fillOpacity}{The opacity of the fill of the map points (0 to 1).}

\item{xlabel}{The label for the variable on the horizontal axis.}

\item{y1label}{The label for the variable on the vertical axis
for the first graph.}

\item{y2label}{The label for the variable on the vertical axis
for the second graph.}

\item{y3label}{The label for the variable on the vertical axis
for the third graph.}

\item{y4label}{The label for the variable on the vertical axis
for the fourth graph.}

\item{showlegend}{Show the plot legend (TRUE) or not (FALSE).}

\item{legendtitle}{The title for the plot legend.}

\item{plottitle}{The title for the plot map.}

\item{spacing}{A value between 0 and 1 for the space between the graphs.}
}
\value{
linked plot map
}
\description{
Create a vehicle trajectory plot map
}
\examples{
# example code
library(crosstalk)
library(dplyr)
data(nds_data)
nds_sf7 <- nds_data |>
  filter(drive==7) |>
  sf::st_as_sf(coords = c("gps_long", "gps_lat"), crs = "WGS84")
nds_sf7_sd <- SharedData$new(nds_sf7)

driveplot(shareddata = nds_sf7_sd,
          x = time_cst,
          y1 = speed_mph,
          y2 = gyro_heading,
          y3 = gps_heading,
          colorvar = gyro_heading,
          maplabel = time_cst,
          colorpalette = "viridis",
          fillOpacity = 1,
          xlabel = "Time",
          y1label = "Speed (mph)",
          y2label = "Gyro Heading (degrees)",
          y3label = "GPS Heading (degrees)")
}
