% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driveplot.R
\name{driveplot}
\alias{driveplot}
\title{Make a stack of companion graphs from shared drive data}
\usage{
driveplot(
  shareddata,
  lng = NULL,
  lat = NULL,
  maplabel = NA,
  colorvar = NULL,
  colorpalette = "#03F",
  fillOpacity = 1,
  time,
  y1,
  y2 = NULL,
  y3 = NULL,
  y4 = NULL,
  timelabel = NULL,
  y1label = NULL,
  y2label = NULL,
  y3label = NULL,
  y4label = NULL,
  showlegend = TRUE,
  legendtitle = NULL,
  plottitle = NULL,
  spacing = 0.05
)
}
\arguments{
\item{shareddata}{a SharedData object containing observations to be plotted}

\item{lng}{the column of shareddata containing longitude (only required if shareddata does not have a geometry column)}

\item{lat}{the column of shareddata containing latitude (only required if shareddata does not have a geometry column)}

\item{maplabel}{an optional label for the map points}

\item{colorvar}{the variable in shareddata to which color should be mapped}

\item{colorpalette}{either a single color (e.g., "red") or one of "viridis", "inferno", "magma", or "plasma"}

\item{fillOpacity}{the opacity of the fill (0 to 1)}

\item{time}{time variable from shareddata to be plotted on the horizontal axis}

\item{y1}{variable from shareddata to be plotted on the vertical axis for the first graph}

\item{y2}{variable from shareddata to be plotted on the vertical axis for the second graph}

\item{y3}{variable from shareddata to be plotted on the vertical axis for the third graph}

\item{y4}{variable from shareddata to be plotted on the vertical axis for the fourth graph}

\item{timelabel}{the label for the time variable on the horizontal axis}

\item{y1label}{the label for the variable on the vertical axis for the first graph}

\item{y2label}{the label for the variable on the vertical axis for the second graph}

\item{y3label}{the label for the variable on the vertical axis for the third graph}

\item{y4label}{the label for the variable on the vertical axis for the fourth graph}

\item{showlegend}{show the plot legend (TRUE) or not (FALSE)}

\item{legendtitle}{the title for the plot legend}

\item{plottitle}{the title for the plot map}

\item{spacing}{value between 0 and 1 for the space between the graphs}
}
\value{
linked plot map
}
\description{
Make a stack of companion graphs from shared drive data
}
