% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driveplot_companion.R
\name{driveplot_companion}
\alias{driveplot_companion}
\title{Make a single plotly scatter plot from shared drive data}
\usage{
driveplot_companion(
  shareddata,
  x,
  y,
  colorvar = NULL,
  colorpalette = "#03F",
  xlabel = NULL,
  ylabel = NULL,
  showlegend = TRUE,
  legendtitle = NULL
)
}
\arguments{
\item{shareddata}{A SharedData object containing observations to be plotted.}

\item{x}{The variable from shareddata to be plotted on the horizontal axis.}

\item{y}{The variable from shareddata to be plotted on the vertical axis.}

\item{colorvar}{The variable in shareddata to which color should be mapped.}

\item{colorpalette}{The color palette for the plot; either a single
color (e.g., "red") or one of "viridis", "inferno", "magma", or "plasma".}

\item{xlabel}{The label for the variable on the horizontal axis.}

\item{ylabel}{The label for the variable on the vertical axis.}

\item{showlegend}{Show the plot legend (TRUE) or not (FALSE).}

\item{legendtitle}{The title for the plot legend.}
}
\value{
A plotly scatterplot.
}
\description{
Make a single plotly scatter plot from shared drive data
}
\examples{
library(crosstalk)
data(drive7)
shared_drive <- SharedData$new(drive7)

# Time series of speed
driveplot_companion(shareddata = shared_drive,
                    x = time_cst,
                    y = speed_mph)

# Color points by direction of car
driveplot_companion(shareddata = shared_drive,
                    x = time_cst,
                    y = speed_mph,
                    colorvar = gyro_heading,
                    colorpalette = "viridis")
}
