% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driveplot_companion.R
\name{driveplot_companion}
\alias{driveplot_companion}
\title{Make a single plotly scatter plot from shared drive data}
\usage{
driveplot_companion(
  shareddata,
  x,
  y,
  colorvar = NULL,
  colorpalette = "#03F",
  xlab = NULL,
  ylab = NULL,
  showlegend = TRUE,
  legendtitle = NULL
)
}
\arguments{
\item{shareddata}{a SharedData object containing observations to be plotted}

\item{x}{variable from shareddata to be plotted on the horizontal axis}

\item{y}{variable from shareddata to be plotted on the vertical axis}

\item{colorvar}{the variable in shareddata to which color should be mapped}

\item{colorpalette}{either a single color (e.g., "red") or one of "viridis", "inferno", "magma", or "plasma"}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{showlegend}{show the plot legend (TRUE) or not (FALSE)}

\item{legendtitle}{the title for the plot legend}
}
\value{
plotly scatterplot
}
\description{
Make a single plotly scatter plot from shared drive data
}
