% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driveplot_companions.R
\name{driveplot_companions}
\alias{driveplot_companions}
\title{Make a stack of companion graphs from shared drive data}
\usage{
driveplot_companions(
  shareddata,
  x,
  y1,
  y2 = NULL,
  y3 = NULL,
  y4 = NULL,
  colorvar = NULL,
  xlabel = NULL,
  y1label = NULL,
  y2label = NULL,
  y3label = NULL,
  y4label = NULL,
  colorpalette = "#03F",
  showlegend = TRUE,
  legendtitle = NULL,
  spacing = 0.05,
  plotheight = "98vh"
)
}
\arguments{
\item{shareddata}{A SharedData object containing observations to be plotted.}

\item{x}{The variable from shareddata to be plotted on the horizontal axis.}

\item{y1}{The variable from shareddata to be plotted on the vertical axis
for the first graph.}

\item{y2}{The variable from shareddata to be plotted on the vertical axis
for the second graph.}

\item{y3}{The variable from shareddata to be plotted on the vertical axis
for the third graph.}

\item{y4}{The variable from shareddata to be plotted on the vertical axis
for the fourth graph.}

\item{colorvar}{The variable in shareddata to which color should be mapped.}

\item{xlabel}{The label for the variable on the horizontal axis.}

\item{y1label}{The label for the variable on the vertical axis
for the first graph.}

\item{y2label}{The label for the variable on the vertical axis
for the second graph.}

\item{y3label}{The label for the variable on the vertical axis
for the third graph.}

\item{y4label}{The label for the variable on the vertical axis
for the fourth graph.}

\item{colorpalette}{The color palette for the plot map; either a single
color (e.g., "red") or one of "viridis", "inferno", "magma", or "plasma".}

\item{showlegend}{Show the plot legend (TRUE) or not (FALSE).}

\item{legendtitle}{The title for the plot legend.}

\item{spacing}{A value between 0 and 1 for the space between the
companion graphs.}

\item{plotheight}{The height of the stack of companion graphs in CSS units,
e.g, "98vh"}
}
\value{
A stack of plotly scatterplots.
}
\description{
Make a stack of companion graphs from shared drive data
}
\examples{
library(crosstalk)
data(drive7)
shared_drive <- SharedData$new(drive7)

# Linked time series of speed, headings (in GPS and gyro), and GPS quality
driveplot_companions(
     shareddata = shared_drive,
     x = time_cst,
     y1 = speed_mph,
     y2 = gyro_heading,
     y3 = gps_heading,
     colorvar = gps_pdop,
     xlabel = "Time",
     y1label = "Speed (mph)",
     y2label = "Gyro Heading (degrees)",
     y3label = "GPS Heading (degrees)",
     colorpalette = "viridis")

}
