% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driveplot_map.R
\name{driveplot_map}
\alias{driveplot_map}
\title{Create a standalone map}
\usage{
driveplot_map(
  shareddata,
  lng = NULL,
  lat = NULL,
  colorvar = NULL,
  label = NA,
  colorpalette = "#03F",
  fillOpacity = 1,
  mapheight = "98vh"
)
}
\arguments{
\item{shareddata}{A SharedData object containing observations to be plotted.}

\item{lng}{The column of shareddata containing longitude
(only required if shareddata does not have a geometry column).}

\item{lat}{The column of shareddata containing latitude
(only required if shareddata does not have a geometry column).}

\item{colorvar}{The column in shareddata to which color should be mapped.}

\item{label}{An optional label for the map points.}

\item{colorpalette}{The color palette for the map; either a single color
(e.g., "red") or one of "viridis", "inferno", "magma", or "plasma".}

\item{fillOpacity}{The opacity of the fill of the map points (0 to 1).}

\item{mapheight}{The height of the map in CSS units, e.g, "98vh".}
}
\value{
A leaflet map.
}
\description{
Create a standalone map
}
\examples{
library(crosstalk)
data(drive7)
shared_drive <- SharedData$new(drive7)

# Basic map of one drive
driveplot_map(shareddata = shared_drive)

# Color drive points by direction of car
driveplot_map(shareddata = shared_drive,
              colorvar = gyro_heading,
              colorpalette = "viridis")
}
